+++ modelsTimescale/streams/stream.js
@@ -29,7 +29,7 @@ module.exports = function (sequelize, DataTypes) {
       allowNull: false
     },
     latitude: {
-      type: DataTypes.DOUBLE,
+      type: DataTypes.REAL,
       allowNull: true,
       validate: {
         isFloat: true,
@@ -44,7 +44,7 @@ module.exports = function (sequelize, DataTypes) {
       }
     },
     longitude: {
-      type: DataTypes.DOUBLE,
+      type: DataTypes.REAL,
       allowNull: true,
       validate: {
         isFloat: true,
@@ -80,8 +80,69 @@ module.exports = function (sequelize, DataTypes) {
     }
   }, {
     paranoid: true,
-    underscored: true
+    timestamps: true,
+    underscored: true,
+    deletedAt: 'deleted_at',
+    hooks: {
+      afterCreate: async (stream, option) => {
+        await updateMinMaxLatLngFromCreate(stream)
+      },
+      afterUpdate: async (stream, option) => {
+        if (stream.projectId !== stream._previousDataValues.projectId) {
+          await updateMinMaxLatLngFromUpdate(stream.projectId)
+          await updateMinMaxLatLngFromUpdate(stream._previousDataValues.projectId)
+        } else if (stream.latitude !== stream._previousDataValues.latitude || stream.longitude !== stream._previousDataValues.longitude) {
+          await updateMinMaxLatLngFromUpdate(stream.projectId)
+        }
+      },
+      afterDestroy: async (stream, option) => {
+        await updateMinMaxLatLngFromUpdate(stream.projectId)
+      }
+    }
   })
+
+  async function updateMinMaxLatLngFromCreate (stream) {
+    const projectId = stream.projectId
+    if (!projectId) {
+      return
+    }
+    const project = await sequelize.models.Project.findByPk(projectId)
+    const update = {}
+    if (project.minLatitude === null || stream.latitude < project.minLatitude) {
+      update.minLatitude = stream.latitude
+    }
+    if (project.maxLatitude === null || stream.latitude > project.maxLatitude) {
+      update.maxLatitude = stream.latitude
+    }
+    if (project.minLongitude === null || stream.longitude < project.minLongitude) {
+      update.minLongitude = stream.longitude
+    }
+    if (project.maxLongitude === null || stream.longitude > project.maxLongitude) {
+      update.maxLongitude = stream.longitude
+    }
+    if (Object.keys(update).length > 0) {
+      await sequelize.models.Project.update(update, { where: { id: projectId } })
+    }
+  }
+
+  async function updateMinMaxLatLngFromUpdate (projectId) {
+    if (!projectId) {
+      return
+    }
+    const update = await sequelize.models.Stream.findAll({
+      plain: true,
+      raw: true,
+      attributes: [
+        [sequelize.fn('min', sequelize.col('latitude')), 'minLatitude'],
+        [sequelize.fn('max', sequelize.col('latitude')), 'maxLatitude'],
+        [sequelize.fn('min', sequelize.col('longitude')), 'minLongitude'],
+        [sequelize.fn('max', sequelize.col('longitude')), 'maxLongitude']
+      ],
+      where: { projectId }
+    })
+    await sequelize.models.Project.update(update, { where: { id: projectId } })
+  }
+
   Stream.associate = function (models) {
     Stream.belongsTo(models.Project, { as: 'project', foreignKey: 'project_id' })
     Stream.belongsTo(models.User, { as: 'created_by', foreignKey: 'created_by_id' })